up/down count

+++ app/Http/Controllers/BasketController.php
@@ -27,17 +27,32 @@ class BasketController extends Controller
         }else{
             $order = Order::find($orderId);
         }
-        $order->products()->attach($productId);
+        if($order->products->contains($productId)){
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            $pivotRow->count++;
+            $pivotRow->update();
+        }else{
+            $order->products()->attach($productId);
+        }
+
         return redirect('bascet');
     }
 
     public function bascetRemove($productId){
         $orderId = session('orderId');
         if(is_null($orderId)){
-            return false;
+            return redirect('bascet');
         }
         $order = Order::find($orderId);
-        $order->products()->detach($productId);
+        if($order->products->contains($productId)){
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            if($pivotRow->count() < 2){
+                $order->products()->detach($productId);
+            }else
+                $pivotRow->count--;
+                $pivotRow->update();
+        }
+        
         return redirect('bascet');
     }
 }
